+++ src/engine/kinematics/index.ts
@@ -6,5 +6,7 @@ export * from './occlusionHeuristics';
 export * from './rulesEngine';
 export * from './referenceDataLoader';
 export * from './dtwCalculator';
+export * from './fatigueDetector';
+